# docs/reservation.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "reservation.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/customer/{customer_id}/reservation": {
      "post": {
        "summary": "Book a new Reservation",
        "description": "Book a new Reservation",
        "operationId": "BookReservation",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/BookReservationResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/BookReservationRequest"
            }
          }
        ],
        "tags": [
          "reservation"
        ]
      }
    },
    "/customer/{customer_id}/reservations": {
      "get": {
        "summary": "List existing reservations per customer ordered by newest.",
        "description": "List existing reservations per customer ordered by newest.",
        "operationId": "GetReservationHistoryPerCustomer",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/GetReservationHistoryPerCustomerResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "customer_id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "uint64"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "uint64"
          }
        ],
        "tags": [
          "reservation"
        ]
      }
    },
    "/reservation/{reservation_id}": {
      "delete": {
        "summary": "Discard an existing reservation",
        "description": "Discard an existing reservation",
        "operationId": "DiscardReservation",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/DiscardReservationResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "reservation_id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "reservation"
        ]
      },
      "put": {
        "summary": "Edit an existing reservation",
        "description": "Edit an existing reservation",
        "operationId": "EditReservation",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/EditReservationResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "reservation_id",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "reservation"
        ]
      }
    }
  },
  "definitions": {
    "BookReservationRequest": {
      "type": "object",
      "properties": {
        "customer_id": {
          "type": "string",
          "format": "int64"
        },
        "reservation": {
          "$ref": "#/definitions/Reservation"
        }
      }
    },
    "BookReservationResponse": {
      "type": "object",
      "properties": {
        "reservation": {
          "$ref": "#/definitions/Reservation"
        },
        "err": {
          "type": "string"
        }
      }
    },
    "DiscardReservationResponse": {
      "type": "object",
      "properties": {
        "err": {
          "type": "string"
        }
      }
    },
    "EditReservationResponse": {
      "type": "object",
      "properties": {
        "reservation": {
          "$ref": "#/definitions/Reservation"
        },
        "err": {
          "type": "string"
        }
      }
    },
    "GetReservationHistoryPerCustomerResponse": {
      "type": "object",
      "properties": {
        "reservations": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Reservation"
          }
        },
        "err": {
          "type": "string"
        }
      }
    },
    "Reservation": {
      "type": "object",
      "properties": {
        "reservation_id": {
          "type": "string",
          "format": "int64"
        },
        "seat_count": {
          "type": "string",
          "format": "int64"
        },
        "start_time": {
          "type": "string"
        },
        "reservation_name": {
          "type": "string"
        },
        "customer_id": {
          "type": "string",
          "format": "int64"
        },
        "phone": {
          "type": "string"
        },
        "comments": {
          "type": "string"
        },
        "created": {
          "type": "string",
          "format": "int64"
        },
        "last_updated": {
          "type": "string",
          "format": "int64"
        }
      }
    }
  }
}